moving all data structures to differents files
