+++ src/Lexer.hs
@@ -5,19 +5,17 @@ module Lexer (
   KeyWord(..),
   TPosition(..),
   Point(..),
-  TokenLine,
   tokenizeProgram,
 ) where
 
 import Utils
+import Point
+import TContext
+import Error
 
 data Token = Token { position :: TPosition , tokenType :: TokenType } deriving (Eq, Show)
 
 data TPosition = TPosition { start :: Point, end :: Point } deriving (Eq, Show)
-
-data Point = Point { row :: Integer, column :: Integer, index :: Integer } deriving (Eq, Show) -- index is the number of the current character respect the complete file
-
-data TContext = TContext { fileName :: String , source :: String , tIndex :: Point } deriving (Eq, Show)
 
 data TokenType =  PToken Puntuation |
                   NToken Int        |
@@ -34,63 +32,13 @@ data KeyWord = Where | Let | In | Import | Backslash |              -- where | l
                Data | Underscore | Class | Type | Instance | Module -- data | _ | class | type | instance | module
                deriving (Eq, Show)
 
-type TokenLine = [TokenType]
-
 ------------------------------
--- Data structure functions --
-
-emptyPoint :: Point
-emptyPoint = Point 0 0 0
-
-updateRow :: Point -> Point
-updateRow Point { row = r, column = c, index = i } = Point (r + 1) 0 i
-
-updateIndex :: Integer -> Point -> Point
-updateIndex n Point { row = r, column = c, index = i } = Point r c (i + n)
-
-updateColumn :: Integer -> Point -> Point
-updateColumn n Point { row = r, column = c, index = i } = Point r (c + n) i
-
-context :: String -> String -> TContext
-context name source = TContext name source emptyPoint
-
-applyIndexContext :: (Point -> Point) -> TContext -> TContext
-applyIndexContext f (TContext {fileName = name, source = s, tIndex = point}) = TContext name s (f point)
-
-incrementRow :: TContext -> TContext
-incrementRow = applyIndexContext updateRow
-
-incrementColumn :: Integer -> TContext -> TContext
-incrementColumn n = applyIndexContext (updateColumn n)
-
-incrementIndex :: Integer -> TContext -> TContext
-incrementIndex n = applyIndexContext (updateIndex n)
-
-pointFromContext :: TContext -> Point
-pointFromContext TContext {fileName = name, source = s, tIndex = point} = point
 
 positionFromContext :: TContext -> TContext -> TPosition
 positionFromContext context nextContext = TPosition (pointFromContext context) (pointFromContext nextContext)
 
 token :: TContext -> TContext -> TokenType -> Token
 token context nextContext tokenType = Token (positionFromContext context nextContext) tokenType
-
-------------------------------
-
--- String aux functions
-
-removeLine :: String -> String
-removeLine "" = ""
-removeLine ('\n':xs) = xs
-removeLine (_:xs) = removeLine xs
-
-nextWord :: String -> String
-nextWord ""                                                      = ""
-nextWord ('-':'-':xs)                                            = ""
-nextWord (' ':xs)                                                = ""
-nextWord ('\n':xs)                                               = "" 
-nextWord (x:xs) | isPuntuation (x:"")                            = ""
-                | otherwise                                      = x : nextWord xs
 
 consumeWord :: TContext -> String -> (String, TContext)
 consumeWord c ""                                 = ("", c)
